LayoutTests/imported/w3c/web-platform-tests/css/css-sizing/contain-intrinsic-size/contain-intrinsic-size-033-expected.txt


PASS .test 1
PASS .test 2
PASS .test 3
PASS .test 4
PASS .test 5
PASS .test 6
PASS .test 7
PASS .test 8
PASS .test 9
PASS .test 10
PASS .test 11
PASS .test 12
PASS .test 13
PASS .test 14
PASS .test 15
PASS .test 16
PASS .test 17
PASS .test 18
FAIL .test 19 assert_equals:
<img src="resources/dice.png" class="test ci-width" data-expected-client-width="10" data-expected-client-height="0">
clientWidth expected 10 but got 0
PASS .test 20
FAIL .test 21 assert_equals:
<img src="resources/dice.png" class="test ci-both" data-expected-client-width="10" data-expected-client-height="20">
clientWidth expected 10 but got 0
FAIL .test 22 assert_equals:
<svg class="test ci-width" data-expected-client-width="10" data-expected-client-height="0"></svg>
clientWidth expected 10 but got 0
PASS .test 23
FAIL .test 24 assert_equals:
<svg class="test ci-both" data-expected-client-width="10" data-expected-client-height="20"></svg>
clientWidth expected 10 but got 0
FAIL .test 25 assert_equals:
<canvas class="test ci-width" data-expected-client-width="10" data-expected-client-height="0"></canvas>
clientWidth expected 10 but got 0
PASS .test 26
FAIL .test 27 assert_equals:
<canvas class="test ci-both" data-expected-client-width="10" data-expected-client-height="20"></canvas>
clientWidth expected 10 but got 0
FAIL .test 28 assert_equals:
<iframe class="test ci-width" data-expected-client-width="10" data-expected-client-height="0"></iframe>
clientWidth expected 10 but got 0
PASS .test 29
FAIL .test 30 assert_equals:
<iframe class="test ci-both" data-expected-client-width="10" data-expected-client-height="20"></iframe>
clientWidth expected 10 but got 0
FAIL .test 31 assert_equals:
<video class="test ci-width" data-expected-client-width="10" data-expected-client-height="0"></video>
clientWidth expected 10 but got 0
PASS .test 32
FAIL .test 33 assert_equals:
<video class="test ci-both" data-expected-client-width="10" data-expected-client-height="20"></video>
clientWidth expected 10 but got 0
PASS .test 34
PASS .test 35
PASS .test 36
PASS .test 37
PASS .test 38
PASS .test 39
PASS .test 40
PASS .test 41
PASS .test 42
